+++ core/src/main/java/com/microsoft/applicationinsights/internal/statsbeat/NetworkStatsbeat.java
@@ -58,7 +58,7 @@ public class NetworkStatsbeat extends BaseStatsbeat {
         }
 
         // send instrumentation as an UTF-8 string
-        String instrumentation = Long.toString(Instrumentations.encode(local.instrumentationList));
+        String instrumentation = String.valueOf(Instrumentations.encode(local.instrumentationList));
 
         if (local.requestSuccessCount.get() != 0) {
             MetricTelemetry requestSuccessCountSt = createStatsbeatTelemetry(REQUEST_SUCCESS_COUNT_METRIC_NAME, local.requestSuccessCount.get());
