+++ core/src/test/java/com/microsoft/applicationinsights/internal/statsbeat/InstrumentationsTest.java
@@ -8,7 +8,7 @@ import java.util.Set;
 
 import static org.junit.Assert.assertEquals;
 
-public class StatsbeatHelperTest {
+public class InstrumentationsTest {
 
     private static final Set<String> instrumentations;
     static {
@@ -20,23 +20,11 @@ public class StatsbeatHelperTest {
 
     private static final long EXPECTED_INSTRUMENTATION = (long)(Math.pow(2, 13) + Math.pow(2, 21) + Math.pow(2, 57)); // Exponents are keys from StatsbeatHelper.INSTRUMENTATION_MAP.)
 
-    private static final Set<Feature> features = Collections.singleton(Feature.JAVA_VENDOR_ZULU);
-
-    private static final long EXPECTED_FEATURE = 2L;
-
     @Test
     public void testEncodeAndDecodeInstrumentations() {
         long longVal = Instrumentations.encode(instrumentations);
         assertEquals(EXPECTED_INSTRUMENTATION, longVal);
         Set<String> result = StatsbeatTestUtils.decodeInstrumentations(longVal);
         assertEquals(instrumentations, result);
-    }
-
-    @Test
-    public void tesEncodeAndDecodeFeature() {
-        long number = Feature.encode(features);
-        assertEquals(EXPECTED_FEATURE, number);
-        Set<Feature> result = StatsbeatTestUtils.decodeFeature(number);
-        assertEquals(features, result);
     }
 }
